requirements.txt
aniso8601==7.0.0
Django==2.2.9
django-filter==2.2.0
graphene==2.1.8
graphene-django==2.8.0
graphql-core==2.2.1
graphql-relay==2.0.1
nose==1.3.7
promise==2.3
pytz==2019.3
Rx==1.6.1
singledispatch==3.4.0.3
six==1.13.0
sqlparse==0.3.0
